Move tracking multiple pages from core to server
The right way to keep track of multiple watches at once is probably
going to depend on what application this crate gets embedded in, so
let's not commit to one design in the core.

Also, stop using `anyhow`; the error-handling situation here is simpler
than I previously thought it was.